Moved functions in pricing.js into location-util.js to be in line with Milo
MM
